Allow MedViewFileWriter to ignore invalid data
When receiving invalid data, MedViewFileWriter will skip the record
instead of raising an error. Because of changes in EmayFileReader
that allow for records to be missing the SpO2 and BPM data, the
writer might get values that can't be converted to an integer, so
revise the error handling to warn and skip, instead of failing.

Signed-off-by: Paul Fagerburg <pfagerburg@gmail.com>