don't skip invalid SpO2/PR when reading
The O2 Insight Pro uses SpO2 == 255 and PR == 65535 when it
cannot get readings (such as when the sensor is off the user's
finger). Rather than transparently skip these records in the
reader class, change the values to None to indicate that there
is a reading for that time, but it is not valid. This behavior
matches what the Emay reader does. The caller can/should filter
out the invalid records.

Signed-off-by: Paul Fagerburg <pfagerburg@gmail.com>